feat: move constants to collection form
